backend/requirements.txt
amqp==5.2.0
asgiref==3.5.0
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==4.4.2
Django==4.0.2
django-celery==3.1.17
django-cors-headers==3.13.0
django-filter==21.1
djangorestframework==3.13.1
gunicorn==21.2.0
idna==3.3
importlib-metadata==4.11.1
isodate==0.6.1
kombu==5.4.2
lxml==5.3.0
Markdown==3.3.6
numpy==2.1.3
packaging==23.1
pillow==10.4.0
platformdirs==4.2.2
proglog==0.1.10
prompt_toolkit==3.0.48
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2021.3
redis==5.2.0
requests==2.27.1
requests-file==2.1.0
requests-toolbelt==1.0.0
setuptools==75.3.0
six==1.16.0
soupsieve==2.6
sqlparse==0.4.2
tqdm==4.67.0
tzdata==2024.2
urllib3==1.26.9
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.0
whitenoise==6.5.0
zeep==4.2.1
zipp==3.7.0
